requirements.txt
Django==1.9.6
django-braces==1.8.1
django-cors-headers==1.1.0
django-extra-fields==0.7
django-filter==0.13.0
django-oauth-toolkit==0.10.0
django-oauth2==1.2
django-oauth2-provider==0.2.6.1
django-push-notifications==1.4.1
djangorestframework==3.1.3
httplib2==0.9.2
Markdown==2.6.6
mock==1.0.1
oauth2==1.5.211
oauth2-provider==0.0
oauthlib==1.0.3
Pillow==3.2.0
psycopg2==2.6.1
pytz==2015.4
requests==2.7.0
shortuuid==0.4.3
six==1.10.0
wheel==0.26.0
